update session creation
