+++ src/app/Http/Controllers/Sessions/Register/Traits/SessionIds.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Http\Controllers\Sessions\Register\Traits;
+
+trait SessionIds
+{
+    protected function getSessionIds(): array
+    {
+        return collect(session('session.sut'))
+            ->pluck('certificate_id')
+            ->filter()
+            ->all();
+    }
+}
