+++ src/database/seeders/test-cases/merchant/authorized-with-account-lookup.yaml
@@ -21,18 +21,22 @@ precondition: |
     - Debit party should be identified by account identifier "msisdn" with the value "+2000009999".
     - Credit party should be identified by account identifier "msisdn" with the value "+33555789123".
 components:
-  - Service Provider
-  - Mojaloop
-  - Mobile Money Operator 1
-  - Mobile Money Operator 2
+  - name: Service Provider
+    slug: service-provider
+  - name: Mobile Money Operator 1
+    slug: mmo-1
+  - name: Mojaloop
+    slug: mojaloop
+  - name: Mobile Money Operator 2
+    slug: mmo-2
 
 test_steps:
   ##------------------------------POST/Transactions: SP -> MMO1------------------------------##
   - path: /transactions
     pattern: '^transactions$'
     method: POST
-    source: Service Provider
-    target: Mobile Money Operator 1
+    source: service-provider
+    target: mmo-1
     api_spec: MM v1.1.2
     trigger:
       amount: '24'
@@ -92,8 +96,8 @@ test_steps:
   - path: /participants/{Type}/{ID}
     pattern: '^participants/MSISDN/[+]2000009999$'
     method: GET
-    source: Mobile Money Operator 1
-    target: Mojaloop
+    source: mmo-1
+    target: mojaloop
     api_spec: Mojaloop v1.0
 
     test_request_scripts:
@@ -122,8 +126,8 @@ test_steps:
   - path: /participants/{Type}/{ID}
     pattern: '^participants/MSISDN/[+]2000009999$'
     method: PUT
-    source: Mojaloop
-    target: Mobile Money Operator 1
+    source: mojaloop
+    target: mmo-1
     api_spec: Mojaloop v1.0
     trigger:
       fspId: 'mmo2fsp'
@@ -157,8 +161,8 @@ test_steps:
   - path: /transactionRequests
     pattern: '^transactionRequests$'
     method: POST
-    source: Mobile Money Operator 1
-    target: Mojaloop
+    source: mmo-1
+    target: mojaloop
     api_spec: Mojaloop v1.0
     trigger:
       amount:
@@ -215,8 +219,8 @@ test_steps:
   - path: /transactionRequests
     pattern: '^transactionRequests$'
     method: POST
-    source: Mojaloop
-    target: Mobile Money Operator 2
+    source: mojaloop
+    target: mmo-2
     api_spec: Mojaloop v1.0
     trigger:
       amount:
@@ -273,8 +277,8 @@ test_steps:
   - path: /transactionRequests/{ID}
     pattern: '^transactionRequests/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
     method: PUT
-    source: Mobile Money Operator 2
-    target: Mojaloop
+    source: mmo-2
+    target: mojaloop
     api_spec: Mojaloop v1.0
 
     test_request_scripts:
@@ -310,8 +314,8 @@ test_steps:
   - path: /transactionRequests/{ID}
     pattern: '^transactionRequests/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
     method: PUT
-    source: Mojaloop
-    target: Mobile Money Operator 1
+    source: mojaloop
+    target: mmo-1
     api_spec: Mojaloop v1.0
 
     test_request_scripts:
@@ -347,8 +351,8 @@ test_steps:
   - path: /quotes
     pattern: '^quotes$'
     method: POST
-    source: Mobile Money Operator 2
-    target: Mojaloop
+    source: mmo-2
+    target: mojaloop
     api_spec: Mojaloop v1.0
     trigger:
       amount:
@@ -418,8 +422,8 @@ test_steps:
   - path: /quotes
     pattern: '^quotes$'
     method: POST
-    source: Mojaloop
-    target: Mobile Money Operator 1
+    source: mojaloop
+    target: mmo-1
     api_spec: Mojaloop v1.0
     trigger:
       amount:
@@ -489,8 +493,8 @@ test_steps:
   - path: /quotes/{ID}
     pattern: '^quotes/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
     method: PUT
-    source: Mobile Money Operator 1
-    target: Mojaloop
+    source: mmo-1
+    target: mojaloop
     api_spec: Mojaloop v1.0
     trigger:
       transferAmount:
@@ -529,8 +533,8 @@ test_steps:
   - path: /quotes/{ID}
     pattern: '^quotes/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
     method: PUT
-    source: Mojaloop
-    target: Mobile Money Operator 2
+    source: mojaloop
+    target: mmo-2
     api_spec: Mojaloop v1.0
     trigger:
       transferAmount:
@@ -569,8 +573,8 @@ test_steps:
   - path: /transfers
     pattern: '^transfers$'
     method: POST
-    source: Mobile Money Operator 2
-    target: Mojaloop
+    source: mmo-2
+    target: mojaloop
     api_spec: Mojaloop v1.0
     trigger:
       amount:
@@ -622,8 +626,8 @@ test_steps:
   - path: /transfers
     pattern: '^transfers$'
     method: POST
-    source: Mojaloop
-    target: Mobile Money Operator 1
+    source: mojaloop
+    target: mmo-1
     api_spec: Mojaloop v1.0
     trigger:
       amount:
@@ -675,8 +679,8 @@ test_steps:
   - path: '{X-Callback-URL}'
     pattern: '(https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|www\.[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\.[^\s]{2,}|https?:\/\/(?:www\.|(?!www))[a-zA-Z0-9]+\.[^\s]{2,}|www\.[a-zA-Z0-9]+\.[^\s]{2,})'
     method: PUT
-    source: Mobile Money Operator 1
-    target: Service Provider
+    source: mmo-1
+    target: service-provider
     trigger:
       amount: '24'
 
@@ -722,8 +726,8 @@ test_steps:
   - path: /transfers/{ID}
     pattern: '^transfers/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
     method: PUT
-    source: Mobile Money Operator 1
-    target: Mojaloop
+    source: mmo-1
+    target: mojaloop
     api_spec: Mojaloop v1.0
 
     test_request_scripts:
@@ -767,8 +771,8 @@ test_steps:
   - path: /transfers/{ID}
     pattern: '^transfers/[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
     method: PUT
-    source: Mojaloop
-    target: Mobile Money Operator 2
+    source: mojaloop
+    target: mmo-2
     api_spec: Mojaloop v1.0
 
     test_request_scripts:
