+++ src/resources/js/layouts/sessions/test-case.vue
@@ -154,9 +154,9 @@
                                                                         getRoute(
                                                                             component.use_encryption,
                                                                             [
+                                                                                component.slug,
+                                                                                connection.slug,
                                                                                 group.id,
-                                                                                component.uuid,
-                                                                                connection.uuid,
                                                                             ],
                                                                             true
                                                                         )
@@ -237,9 +237,9 @@
                                                             getRoute(
                                                                 component.use_encryption,
                                                                 [
+                                                                    component.slug,
+                                                                    connection.slug,
                                                                     session.uuid,
-                                                                    component.uuid,
-                                                                    connection.uuid,
                                                                 ]
                                                             )
                                                         "
