+++ src/resources/js/pages/sessions/register/config.vue
@@ -93,9 +93,9 @@
                                                                 .use_encryption ===
                                                                 '1',
                                                             [
+                                                                sut.slug,
+                                                                connection.slug,
                                                                 group.id,
-                                                                sut.uuid,
-                                                                connection.uuid,
                                                             ],
                                                             true
                                                         )
@@ -154,9 +154,9 @@
                                                     session.sut[sut.id]
                                                         .use_encryption === '1',
                                                     [
+                                                        sut.slug,
+                                                        connection.slug,
                                                         session.info.uuid,
-                                                        sut.uuid,
-                                                        connection.uuid,
                                                     ]
                                                 )
                                             "
@@ -220,7 +220,7 @@
             </div>
             <div class="d-flex justify-content-between">
                 <inertia-link
-                    :href="route('sessions.register.info')"
+                    :href="route('sessions.register.sut')"
                     class="btn btn-outline-primary"
                 >
                     Back
